+++ R/utils.R
@@ -113,3 +113,12 @@ delete_confirmed <- function(confirm, name, type, quote_name=TRUE)
     else utils::askYesNo(msg, FALSE)
     isTRUE(ok)
 }
+
+
+# add a tag on objects created by this package
+add_creator_tag <- function(tags)
+{
+    if(!is.list(tags))
+        tags <- list()
+    utils::modifyList(list(createdBy="AzureR/AzureRMR"), tags)
+}