+++ tests/testthat/test03_rg.R
@@ -31,11 +31,11 @@ test_that("Resource group methods work",
 
     # tagging
     rgnew$set_tags(tag1="value1")
-    expect_identical(rgnew$get_tags(), list(tag1="value1"))
+    expect_identical(rgnew$get_tags(), list(createdBy="AzureR/AzureRMR", tag1="value1"))
     rgnew$set_tags(tag2)
-    expect_identical(rgnew$get_tags(), list(tag1="value1", tag2=""))
+    expect_identical(rgnew$get_tags(), list(createdBy="AzureR/AzureRMR", tag1="value1", tag2=""))
     rgnew$set_tags(tag2=NULL)
-    expect_identical(rgnew$get_tags(), list(tag1="value1"))
+    expect_identical(rgnew$get_tags(), list(createdBy="AzureR/AzureRMR", tag1="value1"))
     rgnew$set_tags(keep_existing=FALSE)
     expect_true(is_empty(rgnew$get_tags()))