+++ NEWS.md
@@ -1,3 +1,8 @@
+# AzureRMR 2.3.3.9000
+
+- Add `createdBy:AzureR/AzureRMR` tag to Azure objects (resource groups, resources and templates) created by this package.
+- Add a `get_tags()` method for templates.
+
 # AzureRMR 2.3.3
 
 - Allow for extra resource type-specific fields beyond those mentioned in the Resource Manager documentation. In particular, virtual machines and managed disks may have a `zones` field containing the availability zones.